+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add `Error::to_compile_error` method
+
 # 0.6.1 - 2019-01-26
 
 * Improve support for `self: Pin<&Self>` and `self: Pin<&mut Self>`